fix(mqtt): 修复发布索引类型及循环变量使用问题
- 将 PublishWithIndex 方法中的 index 参数类型从 int 改为 uint8
- 修正 transformer 中 MQTT 发布循环变量使用错误，使用 id 替代 i
- 确保 MQTT 消息发布时使用正确的索引值

+++ plugin/detection/pkg/mqtt_client.go
@@ -193,7 +193,7 @@ func (m *MQTTClient) Publish(topic string, qos int, payload interface{}) error {
 }
 
 // PublishWithIndex 发布消息，支持索引替换
-func (m *MQTTClient) PublishWithIndex(topicTemplate string, index int, streamPath string, payload interface{}) error {
+func (m *MQTTClient) PublishWithIndex(topicTemplate string, index uint8, streamPath string, payload interface{}) error {
 	if !m.IsConnected() {
 		return fmt.Errorf("MQTT not connected")
 	}
